# backend/internal/api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SwiftInbox External Developer API",
    "version": "1.0.0",
...
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/mailboxes": {
      "post": {
        "tags": ["Mailboxes"],
        "summary": "Create a mailbox",
        "description": "Creates a mailbox for the authenticated developer. The `username` field must be a full mailbox address on the configured mail domain (for example `dev-123@temp.mail.at`).",
        "operationId": "createMailbox",
        "x-usage": { "operation": "mailbox.create", "credits": 1, "category": "mailbox" },
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RequestEmail" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mailbox created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Mailbox" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "get": {
        "tags": ["Mailboxes"],
        "summary": "List mailboxes",
        "description": "Lists all mailboxes owned by the authenticated developer.",
        "operationId": "listMailboxes",
        "x-usage": { "operation": "mailbox.list", "credits": 1, "category": "mailbox" },
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Mailboxes owned by the developer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Mailbox" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/mailboxes/custom": {
      "post": {
        "tags": ["Mailboxes"],
        "summary": "Create a custom mailbox",
        "description": "Creates a mailbox from a username (the local part of the address). The username is sanitized to lowercase alphanumeric characters plus `.`, `-`, and `_` and combined with the configured mail domain. An address that is already in use returns `409 Conflict`.",
        "operationId": "createCustomMailbox",
        "x-usage": { "operation": "mailbox.create", "credits": 1, "category": "mailbox" },
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RequestEmail" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mailbox created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Mailbox" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/api/v1/messages/{id}": {
      "get": {
        "tags": ["Messages"],
        "summary": "Get a message",
        "description": "Fetches a single message body by UUID, including parsed text, HTML, and attachment metadata.",
        "operationId": "getMessage",
        "x-usage": { "operation": "message.get", "credits": 3, "category": "message" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Message UUID.",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "The message.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageDetail" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/api/v1/messages/{id}/attachment/{index}": {
      "get": {
        "tags": ["Attachments"],
        "summary": "Download an attachment",
        "description": "Downloads a single attachment of a message by zero-based index. The response is the raw attachment bytes; its `Content-Type` is the attachment's content type (or `application/octet-stream` when unknown) and `Content-Disposition` is set to `inline` or `attachment`.",
        "operationId": "getAttachment",
        "x-usage": { "operation": "attachment.get", "credits": 2, "category": "message" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Message UUID.",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "index",
            "in": "path",
            "required": true,
            "description": "Zero-based attachment index.",
            "schema": { "type": "integer", "minimum": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "The raw attachment bytes.",
            "content": {
              "application/octet-stream": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "dev_<64 hex chars>",
        "description": "API key sent as a Bearer token in the `Authorization` header (the middleware also accepts the legacy `Authentication` header):\n\n```\nAuthorization: Bearer dev_<64 hex characters>\n```\n\nCreate or manage keys in the developer dashboard (**Api keys** page)."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is missing or malformed (invalid path value, invalid JSON body, wrong domain, invalid expiry, etc.).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, malformed, unknown, or revoked API key.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated developer is not allowed to access the requested resource (for example another developer's mailbox).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "The requested mailbox, message, attachment, webhook, or dead letter does not exist (or does not belong to the developer).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Conflict": {
        "description": "The request conflicts with the current state of the resource (mailbox already exists, mailbox already linked, webhook inactive).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "TooManyRequests": {
        "description": "The developer's usage quota was exceeded (`{\"error\":\"usage quota exceeded\"}`) or the IP rate limit was reached.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "InternalServerError": {
        "description": "An unexpected server error occurred.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        },
        "required": ["error"]
      },
      "RequestEmail": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "For `POST /api/v1/mailboxes` this must be a full mailbox address on the configured mail domain (for example `dev-123@temp.mail.at`). For `POST /api/v1/mailboxes/custom` this is the local part (for example `dev-123`), which is sanitized and combined with the mail domain."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Optional expiry timestamp. Must be in the future; omitted to use the default mailbox lifetime."
          }
        },
        "required": ["username"]
      },
      "Mailbox": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Mailbox id. Present on get/list responses; not returned when a mailbox is created."
          },
          "address": { "type": "string", "description": "Full mailbox address." },
          "createdAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "createdBy": {
            "type": ["string", "null"],
            "format": "uuid",
            "description": "Developer id that created the mailbox, or null for public mailboxes."
          }
        },
        "required": ["address", "createdAt", "expiresAt", "createdBy"]
      },
...
      "AttachmentInfo": {
        "type": "object",
        "properties": {
          "filename": { "type": "string" },
          "contentType": { "type": "string" },
          "size": { "type": "integer", "description": "Attachment size in bytes." },
          "contentId": { "type": "string" },
          "index": { "type": "integer", "description": "Zero-based index used to fetch the attachment." },
          "inline": { "type": "boolean", "description": "True for inline (embedded) parts such as images." }
        },
        "required": ["filename", "contentType", "size", "contentId", "index", "inline"]
      },
      "ParsedData": {
        "type": "object",
        "properties": {
          "subject": { "type": "string" },
          "from": { "type": "string" },
          "text": { "type": "string", "description": "Plain-text body." },
          "html": { "type": "string", "description": "HTML body." },
          "attachments": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AttachmentInfo" }
          }
        },
        "required": ["subject", "from", "text", "html", "attachments"]
      },
      "MessageDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "from": { "type": "string", "example": "sender@example.com" },
          "subject": { "type": "string" },
          "body": {
            "type": "string",
            "description": "Base64-encoded raw MIME message (the Go server serializes the raw message bytes as base64)."
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "mailbox": { "type": "string", "description": "Mailbox address the message belongs to." },
          "parsedData": { "$ref": "#/components/schemas/ParsedData" }
        },
        "required": ["id", "from", "subject", "body", "createdAt", "mailbox", "parsedData"]
      },
...
    }
  }
}