# backend/internal/repository/postgres/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package postgres

import (
	"database/sql"
	"time"

	"github.com/google/uuid"
)

type ApiKeyUsage struct {
	ApiKeyID        uuid.UUID
	Period          time.Time
	ApiRequests     int32
	MailboxRequests int32
	MessageRequests int32
}

type Apikey struct {
	ID          uuid.UUID
	DeveloperID uuid.UUID
	KeyHash     string
	CreatedAt   sql.NullTime
	LastUsedAt  sql.NullTime
	RevokedAt   sql.NullTime
	Name        string
}

type Developer struct {
	ID           uuid.UUID
	Name         string
	Email        string
	CreatedAt    sql.NullTime
	ApiQuota     int32
	MailboxQuota int32
	MessageQuota int32
	PasswordHash string
}

type DeveloperUsage struct {
	DeveloperID      uuid.UUID
	Period           time.Time
	ApiRequests      int32
	MailboxRequests  int32
	MessagesRequests int32
}

type Mailbox struct {
	ID        uuid.UUID
	Address   string
	CreatedAt time.Time
	ExpiresAt time.Time
	CreatedBy uuid.NullUUID
}

type Message struct {
	ID        uuid.UUID
	Sender    string
	Subject   sql.NullString
	MailboxID uuid.UUID
	ExpiresAt time.Time
	CreatedAt time.Time
	Raw       []byte
	Address   string
}

type Webhook struct {
	ID              uuid.UUID
	DeveloperID     uuid.UUID
	Url             string
	SecretEncrypted string
	Events          []string
	IsActive        bool
	CreatedAt       time.Time
	UpdatedAt       time.Time
	Name            string
}

type WebhookDeadLetter struct {
	ID          uuid.UUID
	DeveloperID uuid.UUID
	WebhookID   uuid.UUID
	MailboxID   uuid.NullUUID
	MessageID   uuid.NullUUID
	Event       string
	Url         string
	Reason      string
	Attempts    int32
	CreatedAt   time.Time
	Seen        bool
}

type WebhookMailbox struct {
	WebhookID uuid.UUID
	MailboxID uuid.UUID
	CreatedAt time.Time
}